Create column_correction_draft.py
I think the problem is that the change_state() function allows more than 6 tokens to be added to a column. I also think I found the safety that was supposed to stop that from happening in the original version of minimax, but I think I changed it without catching what it was for because the version that passed all the tests in part 1 had it taken out. Regardless, this is one draft version that was meant to add it back in, I don't think it's a good option & it's causing a lot of logic errors so that